+++ app/src/main/java/handin2/Position.java
@@ -6,19 +6,19 @@ import java.util.ArrayList;
 import javafx.scene.canvas.Canvas;
 
 public class Position {
-    float startScale, factor, zoomedAtX, zoomedAtY;
-    float totalX, totalY;
-    float currentX, currentY;
-    float newScale;
-    float xPosition, yPosition, latPosition, lonPosition;
-    float canvasMaxX, canvasMinX, canvasMaxY, canvasMinY;
+    double startScale, factor, zoomedAtX, zoomedAtY;
+    double totalX, totalY;
+    double currentX, currentY;
+    double newScale;
+    double xPosition, yPosition, latPosition, lonPosition;
+    double canvasMaxX, canvasMinX, canvasMaxY, canvasMinY;
 
-    float canvasWidth, canvasHeight;
-    float maxLat, minLat, maxLon, minLon;
-    float latFactor, lonFactor;
+    double canvasWidth, canvasHeight;
+    double maxLat, minLat, maxLon, minLon;
+    double latFactor, lonFactor;
 
     float[] canvasBoarder;
-    float panY, panX, zoomOut;
+    double panY, panX, zoomOut;
 
     Position(float startScale, Canvas canvas, Model model) {
         this.startScale = startScale;
@@ -26,24 +26,24 @@ public class Position {
         this.zoomedAtX = 0;
         this.zoomedAtY = 0;
 
-        this.canvasHeight = (float) canvas.getHeight();
-        this.canvasWidth = (float) canvas.getWidth();
+        this.canvasHeight = canvas.getHeight();
+        this.canvasWidth = canvas.getWidth();
 
         this.maxLat = model.maxlat;
         this.minLat = model.minlat;
         this.minLon = model.minlon;
 
         // calculates max lon drawn on canvas on start.
-        float latLonFactor = 0.56F;
+        double latLonFactor = 0.56F;
 
-        float latDif = maxLat - minLat;
-        float lonDif = maxLon - minLon;
+        double latDif = maxLat - minLat;
+        double lonDif = maxLon - minLon;
 
-        float lonDifInLat = (maxLon - minLon) * latLonFactor;
-        float canvasFactor = canvasWidth / canvasHeight;
+        double lonDifInLat = (maxLon - minLon) * latLonFactor;
+        double canvasFactor = canvasWidth / canvasHeight;
 
-        float lonNotOnMap = lonDifInLat - (latDif * canvasFactor);
-        float lonDifOnMap = lonDif - (lonNotOnMap / latLonFactor);
+        double lonNotOnMap = lonDifInLat - (latDif * canvasFactor);
+        double lonDifOnMap = lonDif - (lonNotOnMap / latLonFactor);
 
         maxLon = minLon + lonDifOnMap;
 
@@ -56,8 +56,8 @@ public class Position {
 
     // updates the canvas when user expands the size of the window
     public void setnewCanvas(Canvas canvas) {
-        this.canvasHeight = (float) canvas.getHeight();
-        this.canvasWidth = (float) canvas.getWidth();
+        this.canvasHeight = canvas.getHeight();
+        this.canvasWidth = canvas.getWidth();
     }
 
     // updates the factor when zoomed, which is used to calculate mouse- and canvas
@@ -72,7 +72,7 @@ public class Position {
     }
 
     // tells the current mouseposition on the canvas
-    public void setPosition(float x, float y) {
+    public void setPosition(double x, double y) {
         this.currentX = x;
         this.currentY = y;
         setMousePosition();
@@ -94,14 +94,14 @@ public class Position {
     }
 
     // keeps track of how much the canvas has been dragged
-    public void dragged(float x, float y) {
+    public void dragged(Double x, Double y) {
         totalX -= x * factor;
         totalY -= y * factor;
     }
 
     // Updates when zoom. Sets the total x and y to the the total position of the
     // mouse
-    public void mouseZoom(float x, float y) {
+    public void mouseZoom(double x, double y) {
         totalX = xPosition;
         totalY = yPosition;
         zoomedAtX = x;
@@ -118,10 +118,10 @@ public class Position {
 
     // returns the border in lat and lon.
     public float[] getCanvas() {
-        canvasBoarder[0] = maxLat + (latFactor * canvasMinY);
-        canvasBoarder[1] = maxLat + (latFactor * canvasMaxY);
-        canvasBoarder[2] = minLon + (lonFactor * canvasMinX);
-        canvasBoarder[3] = minLon + (lonFactor * canvasMaxX);
+        canvasBoarder[0] = (float) (maxLat + (latFactor * canvasMinY));
+        canvasBoarder[1] = (float) (maxLat + (latFactor * canvasMaxY));
+        canvasBoarder[2] = (float) (minLon + (lonFactor * canvasMinX));
+        canvasBoarder[3] = (float) (minLon + (lonFactor * canvasMaxX));
 
         return canvasBoarder;
     }
@@ -129,11 +129,11 @@ public class Position {
     // used in debugger to show the redbox when freezing frame
     public Way getCanvasOutline() {
         var listnode = new ArrayList<Node>();
-        listnode.add(new Node(canvasBoarder[0], canvasBoarder[2], 0));
-        listnode.add(new Node(canvasBoarder[0], canvasBoarder[3], 2));
-        listnode.add(new Node(canvasBoarder[1], canvasBoarder[3], 3));
-        listnode.add(new Node(canvasBoarder[1], canvasBoarder[2], 6));
-        listnode.add(new Node(canvasBoarder[0], canvasBoarder[2], 0));
+        listnode.add(new Node((float) canvasBoarder[0], (float) canvasBoarder[2], 0));
+        listnode.add(new Node((float) canvasBoarder[0], (float) canvasBoarder[3], 2));
+        listnode.add(new Node((float) canvasBoarder[1], (float) canvasBoarder[3], 3));
+        listnode.add(new Node((float) canvasBoarder[1], (float) canvasBoarder[2], 6));
+        listnode.add(new Node((float) canvasBoarder[0], (float) canvasBoarder[2], 0));
 
         Way way = new Way(listnode);
 
@@ -168,17 +168,17 @@ public class Position {
         panPosition(middleLat, middleLon);
     }
 
-    private void panPosition(Float middleLat, Float middleLon) {
-        Float minLat = canvasBoarder[0];
-        Float maxLat = canvasBoarder[1];
-        Float minLon = canvasBoarder[2];
-        Float maxLon = canvasBoarder[3];
+    private void panPosition(double middleLat, double middleLon) {
+        double minLat = canvasBoarder[0];
+        double maxLat = canvasBoarder[1];
+        double minLon = canvasBoarder[2];
+        double maxLon = canvasBoarder[3];
 
-        float canvasMiddleLat = (minLat + maxLat) / 2;
-        float canvasMiddleLon = (minLon + maxLon) / 2;
+        double canvasMiddleLat = (minLat + maxLat) / 2;
+        double canvasMiddleLon = (minLon + maxLon) / 2;
 
-        float difLat = (canvasMiddleLat - middleLat);
-        float difLon = (canvasMiddleLon - middleLon);
+        double difLat = (canvasMiddleLat - middleLat);
+        double difLon = (canvasMiddleLon - middleLon);
 
         panY = (difLat / latFactor) / factor;
         panX = (difLon / lonFactor) / factor;
